requirements.txt
streamlit
pyspark
kafka-python
pymongo
psycopg2
sqlalchemy
python-dotenv
